Moved files to Android folder.
