+++ android/src/com/mygdx/game/Simulation/BodyLimb.java
@@ -1,4 +1,4 @@
-package com.mygdx.game;
+package com.mygdx.game.Simulation;
 
 import com.badlogic.gdx.graphics.VertexAttributes;
 import com.badlogic.gdx.graphics.g3d.Material;
