+++ android/src/com/mygdx/game/Simulation/HelperClass.java
@@ -1,4 +1,4 @@
-package com.mygdx.game;
+package com.mygdx.game.Simulation;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
